+++ packages/core/CHANGELOG.md
@@ -1,8 +1,20 @@
 # Changelog
 
-## [2.2.21] - 2026-07-30
+## [2.2.24] - 2026-07-30
 
 - Add new Star Rating input validation type
+
+## [2.2.23] - 2026-08-04
+
+- Update address complete validation
+
+## [2.2.22] - 2026-07-31
+
+- Move address complete validation to helper
+
+## [2.2.21] - 2026-07-31
+
+- Update validation for address complete
 
 ## [2.2.20] - 2026-07-30